refactor(opal): move Toast to layouts
