+++ web/lib/opal/src/layouts/index.ts
@@ -80,3 +80,18 @@ export type {
 
 /* TagList */
 export { TagList, type TagListProps } from "@opal/layouts/tag-list/components";
+
+/* Toast */
+export {
+  ToastProvider,
+  type ToastProviderProps,
+} from "@opal/layouts/toast/components";
+export {
+  toast,
+  useToast,
+  useToastFromQuery,
+  MAX_VISIBLE_TOASTS,
+  type Toast,
+  type ToastLevel,
+  type ToastOptions,
+} from "@opal/layouts/toast/store";
