+++ web/lib/opal/src/layouts/toast/components.tsx
@@ -10,7 +10,7 @@ import {
   toastStore,
   type Toast,
   type ToastLevel,
-} from "@opal/components/toast/store";
+} from "@opal/layouts/toast/store";
 
 const ANIMATION_DURATION = 200; // matches tailwind fade-out-scale (0.2s)
 const MAX_TOAST_MESSAGE_LENGTH = 150;
